refactor: delete package which contains single file
